加入进程池异步并发处理训练模型请求

+++ libs/email.py
@@ -18,7 +18,7 @@ def send_email():
         result = check_alarm(single['site'], single['pollution'], single['value'])
         if (result['state'] == True):
             receiver = single['email']
-            percent = '%.2f%%' % ((int(result['db_value']) - int(single['value'])) / int(single['value']) * 100)
+            percent = '%.2f%%' % ((float(result['db_value']) - float(single['value'])) / float(single['value']) * 100)
             subject = "@AirNet污染物预警！监测到" +"观测点："+single['site']+ "的污染物："+single['pollution'] + "已超标"+percent
             mail_msg = "<p>您好，这里是AirNet。</p><p>我们已监测到您设定的污染物：【" + single['pollution'] + "】已超过预警值：【" \
                        + str(single['value']) + "】</p><p>此次监测时间：【"+result['db_time']+"】，监测值为：【"+str(result['db_value'])+"】</p><p>感谢您对AirNet的支持</p>"
